requirements.txt
streamlit
langchain
python-dotenv
pandas    
langchain_community   
pypdf
groq
langchain_groq